backend/requirements.txt
passlib
motor
requests
selenium
python-jose
PyJWT